feat: Add scheduled_for variable to createSchedule function
This commit adds a new variable called `scheduled_for` to the `createSchedule` function in the `create-schedule.ts` file. Currently, it is assigned a placeholder value of "--". This variable will be used to store the calculated next scheduled time based on the provided cron expression. The actual calculation logic will be implemented later.

Because of this issue: ` Error: Constraint error, got value 28 expected range 0-23`

Co-authored-by: [Jackson Kasi] <[jacksonkasipeacock@gmail.com]>